refactor: elevate render_attributes
