+++ orbiter/objects/operators/empty.py
@@ -1,8 +1,8 @@
 from typing import Literal
 
-from orbiter.objects import ImportList
+from orbiter.objects import ImportList, RenderAttributes
 from orbiter.objects.requirement import OrbiterRequirement
-from orbiter.objects.task import OrbiterOperator, RenderAttributes
+from orbiter.objects.task import OrbiterOperator
 
 
 class OrbiterEmptyOperator(OrbiterOperator):
