+++ orbiter/objects/operators/ssh.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 from typing import Dict, Literal
 
-from orbiter.objects import ImportList
+from orbiter.objects import ImportList, RenderAttributes
 from orbiter.objects.requirement import OrbiterRequirement
-from orbiter.objects.task import OrbiterOperator, RenderAttributes
+from orbiter.objects.task import OrbiterOperator
 
 
 class OrbiterSSHOperator(OrbiterOperator):
